Refactor PlayerHop into a manager script and an object script to make individual objects bounce
