+++ Assets/Scripts/ManagerScripts/PlayerHopManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 40c903874530b3f4994bbf86e2ba9689
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
