+++ Assets/Scripts/ManagerScripts/RespawnManager.cs
@@ -41,7 +41,7 @@ public class RespawnManager : MonoBehaviour
         _inputIndexSBA = SingleButtonAction.Current.GetInputIndex();
         _inputIndexPSA = PlayerSideAction.Current.GetInputIndex();
         _inputIndexRightPSA = PlayerSideAction.Current.GetInputIndexRight();
-        _hopIndex = PlayerHop.Current.GetHopIndex();
+        _hopIndex = PlayerHopManager.Current.GetHopIndex();
     }
 
     public IEnumerator RespawnPlayer(float respawnClipLength)
@@ -61,7 +61,7 @@ public class RespawnManager : MonoBehaviour
         SingleButtonAction.Current.SetInputIndex(_inputIndexSBA);
         PlayerSideAction.Current.SetInputIndex(_inputIndexPSA);
         PlayerSideAction.Current.SetInputIndexRight(_inputIndexRightPSA);
-        PlayerHop.Current.SetHopIndex(_hopIndex);
+        PlayerHopManager.Current.SetHopIndex(_hopIndex);
 
         // Reset music related values
         MusicPlayer.Current.Resume();
