+++ Assets/Scripts/ObjectScripts/PlayerHop.cs
@@ -1,35 +1,17 @@
-using System;
+﻿using System;
 using UnityEngine;
 
 public class PlayerHop : MonoBehaviour
 {
-    public static PlayerHop Current;
-
     private Rigidbody _rb;
     private Vector3 _startPos;
     private bool _hopping;
-    private int _hopIndex;
-
-    private void Awake()
-    {
-        Current = this;
-    }
-
     private void Start()
     {
         _rb = GetComponent<Rigidbody>();
+        PlayerHopManager.Current.HopEvent += Hop;
         _startPos = transform.position;
         _hopping = false;
-    }
-
-    public int GetHopIndex()
-    {
-        return _hopIndex;
-    }
-
-    public void SetHopIndex(int hopI)
-    {
-        _hopIndex = hopI;
     }
 
     private void Update()
@@ -40,26 +22,19 @@ public class PlayerHop : MonoBehaviour
             _rb.velocity = new Vector3(0, 0, 0);
             _hopping = false;
         }
-        if (Math.Abs(MusicPlayer.Current.GetAudioSourceTime() - SingleButtonAction.Current.GetNextTimestamp(_hopIndex)) < 0.1f)
-        {
-            Hop();
-        }
     }
 
     private void FixedUpdate()
     {
         _rb.AddForce(Vector3.down * ((PlayerMovement.Current.jumpingGravity + 5.5f) * _rb.mass));
-
     }
-
+    
     private void Hop()
     {
         if (_hopping == false)
         {
             _hopping = true;
             _rb.AddForce(transform.up * PlayerMovement.Current.maxJumpForce, ForceMode.Impulse);
-            _hopIndex++;
         }
     }
-    
 }
