+++ Assets/Scripts/ObjectScripts/PlayerHop.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 40c903874530b3f4994bbf86e2ba9689
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 375d8d89f94b439b85be3ef7cf4ab6c2
+timeCreated: 1670181704
